requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.10.1
mongomock==4.0.0
gunicorn==22.0.0
pytest==8.3.4
pyyaml==6.0.2
